Limit hardware profiles to machineTypes with disks
Some machineTypes do not come with any disks, and will error if
attempted to boot one of these types without explicitly adding disks.
Having the compute service create a disk doesn't work as that means you
can't choose an image (they are mutually exclusive). So simplest option
to reduce confusion is to not allow the compute service to select
machineTypes without a disk.