Resolve the project ID to a name
This makes the @UserProject Supplier<String> always return a project
name, instead of the project ID number. Some resources always return the
project name, which can result in some objects using the ID and others
using the name, and this inconsistency causes problems (e.g. when trying
to look up locations in a hashtable). For consistency, standardize on
using the project name.