Fix GCE nodes n>1 ignoring inboundPort
The inboundPort settings of the first node in the group dictated the
firewall configuration. Subsequent nodes added to the group had their
inboundPort settings ignored. This commit makes GCE create a firewall
for each port targetting a specific tag, and tags each instance with the
details of its inboundPort settings.